pkp/pkp-lib#2163 Work on status tracking of review rounds and assignments.
ReviewRound and ReviewAssigment objects get defined statuses. This commit
extends those statuses, makes sure they're recalculated at the appropriate
times, and moves that calculation into the objects instead of grids.

IN the process, it removes the all-reviews-are-in task. (This will get
surfaced as a review round status in the submissions list.) It also
removes the now obsolete replaced state for ReviewAssignments.

Whenever a ReviewAssignment is updated, the ReviewRound status should get
updated automatically. So the _updateReviewRoundStatus() helper method in
PKPReviewGridHandler has been removed. (A similarly named method in
EditorDecisions is still in play, as it manually overrides the status when
an EditorDecision is invoked.)

Also fixes a couple ofbugs with the reviewer grid and round status not
updating in response to actions.