pkp/pkp-lib#2163 Fix db error when adding a reviewer.
This error was introduced during a rebase. When the replaced review
assignment flag was removed, a manual rebase had to be done and the
number of arguments in the db statement got out-of-sync. That rebase
fix was https://github.com/NateWr/pkp-lib/commit/58e40d9f9dda5d7c806685100c6b894c41ccfd53